requirements.txt
langchain_community 
beautifulsoup4
duckduckgo-search
langchain==0.3.4
youtube-search-python
youtube-transcript-api
streamlit
googletrans
cython>=0.29.21
numpy>=1.10.4
scipy>=0.13.3
pbr
six
future
praw
langchain_groq
pydantic
tavily-python
faiss-cpu
sentence_transformers
pinecone